- Added bind_back() - moved contents of invoke.h into functional.h - invoke.h just includes functional.h - added tests
